Added Linux 64 bit fixes to the ui code.
git-svn-id: svn://0x263a.de/tekuum_src@214 3ae3452d-76a7-4814-9b33-b51e13292e42

+++ ui/Window.h
@@ -105,8 +105,10 @@ typedef enum {
 } wexpRegister_t;
 
 typedef struct {
-	wexpOpType_t opType;	
-	int	a, b, c, d;
+	wexpOpType_t opType;
+	// Techyon RB: 64 bit fixes, changed int to intptr_t
+	intptr_t	a, b, c, d;
+	// Techyon END
 } wexpOp_t;
 
 struct idRegEntry {
